+++ src/deluge/gui/menu_item/source/patched_param.h
@@ -22,6 +22,6 @@ namespace deluge::gui::menu_item::source {
 class PatchedParam : public menu_item::patched_param::Integer {
 public:
 	using Integer::Integer;
-	uint8_t getP() override { return menu_item::PatchedParam::getP() + soundEditor.currentSourceIndex; }
+	uint8_t getP() override { return menu_item::PatchedParam::getP() + soundEditor.currentPatchIndex; }
 };
 } // namespace deluge::gui::menu_item::source